requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Machine Learning & AI
scikit-learn==1.3.2
TA-Lib==0.4.28
xgboost==2.0.2
torch==2.1.2
transformers==4.36.2
textblob==0.17.1

# Technical Analysis
ta==0.10.2
pandas-ta==0.3.14b

# Financial Data
yfinance==0.2.28
ccxt==4.1.77

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1

# Monitoring & Observability
prometheus-client==0.19.0
psutil==5.9.6
structlog==23.2.0

# Security & Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# WebSocket & Real-time
websockets==12.0
aiofiles==23.2.1

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Production & Deployment
gunicorn==21.2.0
supervisor==4.2.5

# Backup & Storage
boto3==1.34.0
minio==7.2.0

# Time & Date
python-dateutil==2.8.2
pytz==2023.3

# Utilities
click==8.1.7
tqdm==4.66.1
backoff==2.2.1
tenacity==8.2.3

# Development Tools
jupyter==1.0.0
ipython==8.18.1
notebook==7.0.6

# Security Scanning
bandit==1.7.5
safety==2.3.5

# Performance & Profiling
memory-profiler==0.61.0
line-profiler==4.1.2

# Logging & Debugging
loguru==0.7.2
rich==13.7.0

# API Documentation
swagger-ui-bundle==0.0.9

# Data Validation
cerberus==1.3.5

# Caching
cachetools==5.3.2

# Rate Limiting
slowapi==0.1.9

# Health Checks
healthcheck==1.3.5

# Metrics & Monitoring
prometheus-client==0.19.0
grafana-api==1.0.3

# Message Queue
pika==1.3.2

# File Processing
openpyxl==3.1.2
xlrd==2.0.1

# Image Processing (for charts)
pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Network & Connectivity
aiofiles==23.2.1
aiohttp-cors==0.7.0

# Compression
gzip==1.0.0

# Encryption
cryptography==41.0.8

# Environment Management
virtualenv==20.25.0

# Package Management
pip-tools==7.3.0

# Development Server
watchdog==3.0.0

# API Testing
pytest-httpx==0.25.0

# Database Testing
pytest-postgresql==4.1.5

# Mock Testing
pytest-mock==3.12.0

# Coverage
coverage==7.3.2

# Type Checking
types-requests==2.31.0.10
types-PyYAML==6.0.12.12

# Production Dependencies
gunicorn==21.2.0
uvloop==0.19.0

# Monitoring Dependencies
prometheus-client==0.19.0
grafana-api==1.0.3

# Backup Dependencies
boto3==1.34.0
minio==7.2.0

# Security Dependencies
cryptography==41.0.8
passlib[bcrypt]==1.7.4

# Performance Dependencies
uvloop==0.19.0
orjson==3.9.10

# Development Dependencies
jupyter==1.0.0
ipython==8.18.1
notebook==7.0.6
